Universal Scheduler/add_crud_functionality.py
#!/usr/bin/env python3
"""Add full CRUD functionality to all sections of the system."""

import requests
import json
from datetime import date, datetime

API_BASE = "http://127.0.0.1:8000/api"

def add_delete_endpoints():
    """Add DELETE endpoints to the workflow API for clearing sections."""
    
    # This will show what endpoints we need to add
    endpoints_needed = [
        # Academic Year Management
        "DELETE /api/workflow/academic-years/{year}",
        "DELETE /api/workflow/academic-years/{year}/residents",
        "DELETE /api/workflow/academic-years/{year}/preferences",
        "DELETE /api/workflow/academic-years/{year}/holidays",
        "DELETE /api/workflow/academic-years/{year}/historical-imports",
        
        # Residents
        "DELETE /api/residents/{id}",
        "DELETE /api/residents",  # Clear all
        
        # Holiday Management
        "DELETE /api/workflow/holiday-definitions/{id}",
        "DELETE /api/workflow/academic-years/{year}/dynamic-holidays/{id}",
        "DELETE /api/workflow/academic-years/{year}/holiday-periods/{id}",
        
        # Preferences and Time Off
        "DELETE /api/residents/{id}/preferences",
        "DELETE /api/residents/{id}/time-off/{request_id}",
        
        # System Reset
        "DELETE /api/system/reset",  # Clear everything
        "POST /api/system/save-state",  # Save current state
        "POST /api/system/restore-state",  # Restore saved state
    ]
    
    print("🔧 CRUD Endpoints Needed:")
    for endpoint in endpoints_needed:
        print(f"   {endpoint}")
    
    return endpoints_needed

def test_current_edit_capabilities():
    """Test what edit capabilities currently exist."""
    print("\n🧪 Testing Current Edit Capabilities:")
    
    # Test if we can get residents for editing
    try:
        response = requests.get(f"{API_BASE}/residents")
        if response.status_code == 200:
            residents = response.json()
            print(f"   ✅ Can list residents: {len(residents)} found")
        else:
            print(f"   ❌ Cannot list residents: {response.status_code}")
    except:
        print(f"   ❌ Error accessing residents")
    
    # Test holiday definitions
    try:
        response = requests.get(f"{API_BASE}/workflow/holiday-definitions")
        if response.status_code == 200:
            holidays = response.json()
            print(f"   ✅ Can list holiday definitions: {len(holidays)} found")
        else:
            print(f"   ❌ Cannot list holiday definitions")
    except:
        print(f"   ❌ Error accessing holiday definitions")
    
    # Test academic year workflow
    try:
        response = requests.get(f"{API_BASE}/workflow/academic-years")
        if response.status_code == 200:
            workflows = response.json()
            print(f"   ✅ Can list academic years: {len(workflows)} found")
        else:
            print(f"   ❌ Cannot list academic years")
    except:
        print(f"   ❌ Error accessing academic years")

def main():
    print("🔧 CRUD FUNCTIONALITY ANALYSIS")
    print("="*50)
    
    # Show what endpoints we need
    add_delete_endpoints()
    
    # Test current capabilities
    test_current_edit_capabilities()
    
    print(f"\n📝 IMPLEMENTATION PLAN:")
    print(f"   1. Add DELETE endpoints for each section")
    print(f"   2. Add PUT/PATCH endpoints for editing")
    print(f"   3. Add bulk operations (clear all, save state)")
    print(f"   4. Update Streamlit UI with edit/delete buttons")
    print(f"   5. Add confirmation dialogs")
    print(f"   6. Add save/restore functionality")

if __name__ == "__main__":
    main()