Universal Scheduler/requirements.txt
# Web Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit-option-menu>=0.3.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.7
redis>=5.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
openpyxl>=3.1.0

# Optimization
ortools>=9.7.0
python-constraint>=1.4.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.0

# Legacy dependencies (to be removed)
python-docx>=0.8.11
PyPDF2>=3.0.0
spacy>=3.6.0
nltk>=3.8.0